[update] EXHashTag 컴포넌트 로직 분리   #160
